@@ -1,5 +1,4 @@
 import { TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
 import { of } from 'rxjs';
 import { OktaAuthStateService, OKTA_AUTH } from '@okta/okta-angular';
@@ -11,14 +10,11 @@ describe('AppComponent', () => {
     })
   });
 
-  const authSpy = jasmine.createSpyObj('OktaAuth', ['login']);
+  const authSpy = jasmine.createSpyObj('OktaAuth', ['signInWithRedirect']);
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
-        RouterTestingModule
-      ],
-      declarations: [
         AppComponent
       ],
       providers: [
@@ -34,9 +30,11 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'okta-angular-quickstart'`, () => {
+  it(`should call Okta's login method in the sign in method`, async () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('okta-angular-quickstart');
-  });
+    await app.signIn().then(() => {
+      expect(authSpy.signInWithRedirect).toHaveBeenCalled();
+    });
+  })
 });
